chore: run isort and black.

+++ lal/src/preprocessing.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import json
 import logging
 from typing import Any, List, TypedDict
 
@@ -7,7 +8,6 @@ import pandas as pd
 import psutil
 from datasets import Dataset
 from transformers import PreTrainedTokenizerBase, PreTrainedTokenizerFast
-import json
 
 from ..conf.config import Composer
 
@@ -53,17 +53,18 @@ def add_prompt_name_group(df_1: pd.DataFrame, df_2: pd.DataFrame) -> pd.DataFram
     merged_df = pd.merge(df_1, df_2, on="essay_id", how="left")
     return merged_df[df_1.columns.tolist() + ["prompt_name"]].reset_index(drop=True)
 
+
 def merge_topic_info_to_df(df: pd.DataFrame, train_topic_filepath: str, topics_map_path: str) -> pd.DataFrame:
     train_topic_df = pd.read_csv(train_topic_filepath)
-    train_topic_df = train_topic_df[['essay_id', 'topics']]
+    train_topic_df = train_topic_df[["essay_id", "topics"]]
     merge_df_and_train_topic = pd.merge(df, train_topic_df, on="essay_id", how="left")
-    with open(topics_map_path, 'r') as file:
+    with open(topics_map_path, "r") as file:
         topics_dict = json.load(file)
 
-    topics_df = pd.DataFrame(list(topics_dict.items()), columns=['topics', 'description'])
-    topics_df['topics'] = topics_df['topics'].astype(int)
-    final_merged_df = pd.merge(merge_df_and_train_topic, topics_df, on='topics', how='left')
-    final_merged_df = final_merged_df.drop(columns=['topics'])
+    topics_df = pd.DataFrame(list(topics_dict.items()), columns=["topics", "description"])
+    topics_df["topics"] = topics_df["topics"].astype(int)
+    final_merged_df = pd.merge(merge_df_and_train_topic, topics_df, on="topics", how="left")
+    final_merged_df = final_merged_df.drop(columns=["topics"])
     return final_merged_df
